requirements/requirements.txt
matplotlib==3.4.3
numpy==1.21.1
pandas==1.3.1
nltk==3.6.3
scikit-learn==0.24.2
tensorflow==2.5.0
requests==2.26.0
beautifulsoup4==4.10.0
mysql-connector-python==8.0.26
